packet::ether

Struct Packet

pub struct Packet<B> { /* private fields */ }

Ethernet frame parser.

Implementations

impl<B: AsRef<[u8]>> Packet<B>

pub fn unchecked(buffer: B) -> Packet<B>

Create an Ethernet frame without checking the buffer.

pub fn new(buffer: B) -> Result<Packet<B>>

Parse an Ethernet frame, checking the buffer contents are correct.

impl<B: AsRef<[u8]>> Packet<B>

pub fn to_owned(&self) -> Packet<Vec<u8>>

Convert the packet to its owned version.

Notes

It would be nice if ToOwned could be implemented, but Packet already implements Clone and the impl would conflict.

impl<B: AsRef<[u8]>> Packet<B>

pub fn destination(&self) -> HwAddr

MAC address for the destination.

pub fn source(&self) -> HwAddr

MAC address for the source.

pub fn protocol(&self) -> Protocol

Protocol of the inner packet.

impl<B: AsRef<[u8]> + AsMut<[u8]>> Packet<B>

pub fn set_destination(&mut self, value: HwAddr) -> Result<&mut Self>

Destination MAC address.

pub fn set_source(&mut self, value: HwAddr) -> Result<&mut Self>

Source MAC address.

pub fn set_protocol(&mut self, value: Protocol) -> Result<&mut Self>

Inner protocol.